BA-1093 - Fix proposal submission deadline

+++ modules/proposals/client/controllers/ProposalEditSWUController.ts
@@ -7,7 +7,7 @@ import moment from 'moment-timezone';
 import { Settings } from 'tinymce';
 import { ICapabilityResource } from '../../../capabilities/client/services/CapabilitiesService';
 import { ICapabilitySkillResource } from '../../../capabilities/client/services/CapabilitiesSkillsService';
-import { IOpportunityResource } from '../../../opportunities/client/services/OpportunitiesService';
+import { IOpportunitiesService, IOpportunityResource } from '../../../opportunities/client/services/OpportunitiesService';
 import { IOrgCommonService } from '../../../orgs/client/services/OrgCommonService';
 import { IOrgResource } from '../../../orgs/client/services/OrgService';
 import { IAuthenticationService } from '../../../users/client/services/AuthenticationService';
@@ -29,7 +29,8 @@ export default class ProposalEditSWUController {
 		'TinyMceConfiguration',
 		'resources',
 		'$window',
-		'OrgCommonService'
+		'OrgCommonService',
+		'OpportunitiesService'
 	];
 
 	public title: string;
@@ -65,7 +66,8 @@ export default class ProposalEditSWUController {
 		public TinyMceConfiguration: Settings,
 		public resources: any,
 		private $window: IWindowService,
-		private OrgCommonService: IOrgCommonService
+		private OrgCommonService: IOrgCommonService,
+		private OpportunitiesService: IOpportunitiesService
 	) {
 		this.user = this.AuthenticationService.user;
 		this.activeTab = 1;
@@ -164,6 +166,10 @@ export default class ProposalEditSWUController {
 
 	// perform the save, both user info and proposal info
 	public async save(successMessage?: string): Promise<void> {
+		if (!(await this.checkDeadline())) {
+			return;
+		}
+
 		if (!successMessage) {
 			successMessage = 'Changes saved';
 		}
@@ -189,13 +195,19 @@ export default class ProposalEditSWUController {
 		}
 	}
 
-	public withdraw() {
-		this.proposal.status = 'Draft';
-		this.save('Your proposal has been withdrawn');
+	public async withdraw(): Promise<void> {
+		if (await this.checkDeadline()) {
+			this.proposal.status = 'Draft';
+			this.save('Your proposal has been withdrawn');
+		}
 	}
 
 	// submit the proposal
 	public async submit(): Promise<void> {
+		if (!(await this.checkDeadline())) {
+			return;
+		}
+
 		// validate the phase and total amounts
 		if (!this.validateAllAmounts()) {
 			this.Notification.error({
@@ -242,6 +254,10 @@ export default class ProposalEditSWUController {
 
 	// delete the proposal with confirmation
 	public async delete(): Promise<void> {
+		if (!(await this.checkDeadline())) {
+			return;
+		}
+
 		const question = 'Are you sure you want to permanently delete your proposal?';
 		const choice = await this.ask.yesNo(question);
 		if (choice) {
@@ -259,6 +275,10 @@ export default class ProposalEditSWUController {
 
 	// upload documents
 	public async upload(file: File): Promise<void> {
+		if (!(await this.checkDeadline())) {
+			return;
+		}
+
 		if (!file) {
 			return;
 		}
@@ -292,6 +312,10 @@ export default class ProposalEditSWUController {
 	}
 
 	public async deletefile(fileId: string): Promise<void> {
+		if (!(await this.checkDeadline())) {
+			return;
+		}
+
 		const updatedProposal = await this.ProposalService.removeDoc({
 			proposalId: this.proposal._id,
 			documentId: fileId
@@ -321,6 +345,20 @@ export default class ProposalEditSWUController {
 
 	public toggleGuidance(index: number): void {
 		this.opportunity.teamQuestions[index].showGuidance = !this.opportunity.teamQuestions[index].showGuidance;
+	}
+
+	private async checkDeadline(): Promise<boolean> {
+		// Check with server to ensure deadline hasn't passed
+		const response = await this.OpportunitiesService.getDeadlineStatus({ opportunityId: this.opportunity._id }).$promise;
+		if (response.deadlineStatus === 'CLOSED') {
+			this.Notification.error({
+				title: 'Error',
+				message: '<i class="fas fa-exclamation-triangle"></i> Deadline has passed!'
+			});
+			return false;
+		} else {
+			return true;
+		}
 	}
 
 	private isTeamCapable(): boolean {
